Tweaks
I added tag data to the similar_content endpoint; also, id is now a
column for tags and contents in the admin interface. This helps with
understanding the quality of the tag similarity metric - is it returning
results that I would personally consider relevant? Where does it get
tripped up?