Select and deselect tags
in a natural way, by simply clicking on any instance of the tag. The
selected tags are shown in green to visually distinguish them from
unselected tags.

I also cleaned up the animations, tightening the flow by making multiple
effects execute in parallel. The result is a faster feel to the UI.